Rebuild dkfsserver, modify mapper.

See merge request rapserver/smarthelper!628

+++ dkfsweb/src/main/java/com/rpa/web/service/IUserVipService.java
@@ -32,5 +32,5 @@ public interface IUserVipService {
      * @param userId
      * @return
      */
-    DTPageInfo<UserVipDetailsVO> queryDetails(int draw, int pageNum, int pageSize, int userId);
+    DTPageInfo<UserVipDetailsVO> queryDetails(int draw, int pageNum, int pageSize, long userId);
 }
